+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
 
 * Specify precision and scale for ``DECIMAL`` if the dialect is ``mssql``, ``mysql`` or ``oracle``.
 * Set length of ``VARCHAR`` to ``1`` even if maximum length is ``0`` to satisfy MySQL.
+* Set type to ``TEXT`` if maximum length is greater than 21,844 to satisfy MySQL.
 * Add ``chunk_size`` option to :meth:`.Table.to_sql` to write rows in batches.
 * Add ``unique_constraint`` option to :meth:`.Table.to_sql` to include in a UNIQUE constraint.